+++ src/algorithms/flow/mod.rs
@@ -14,8 +14,10 @@ pub(crate) mod edge_disjoint_paths;
 pub(crate) mod max_flow;
 pub(crate) mod st_edge_connectivity;
 pub(crate) mod st_mincut;
+pub(crate) mod st_vertex_connectivity;
 
 pub use edge_disjoint_paths::edge_disjoint_paths;
 pub use max_flow::max_flow_value;
 pub use st_edge_connectivity::st_edge_connectivity;
 pub use st_mincut::st_mincut_value;
+pub use st_vertex_connectivity::{VconnNei, st_vertex_connectivity};