+++ src/lib.rs
@@ -137,6 +137,7 @@ pub use crate::algorithms::flow::edge_disjoint_paths::edge_disjoint_paths;
 pub use crate::algorithms::flow::max_flow::max_flow_value;
 pub use crate::algorithms::flow::st_edge_connectivity::st_edge_connectivity;
 pub use crate::algorithms::flow::st_mincut::st_mincut_value;
+pub use crate::algorithms::flow::st_vertex_connectivity::{VconnNei, st_vertex_connectivity};
 pub use crate::algorithms::games::barabasi::barabasi_game_bag;
 pub use crate::algorithms::games::barabasi_aging::barabasi_aging_game;
 pub use crate::algorithms::games::barabasi_psumtree::{